requirements.txt
alembic==1.8.1
APScheduler==3.10.4
greenlet==3.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.1.1
paho-mqtt==1.6.1
python-dotenv==0.21.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==1.4.44
tzdata==2023.3
tzlocal==5.2
